+++ src/renderer/components/dialogs/Settings-Advanced.tsx
@@ -3,6 +3,7 @@ import React from 'react'
 import { SettingsStoreState } from '../../stores/settings'
 import { RenderDeltaSwitch2Type, RenderDTSettingSwitchType } from './Settings'
 import SettingsEncryption from './Settings-Encryption'
+import { SettingsExperimentalFeatures } from './Settings-ExperimentalFeatures'
 import SettingsImapFolderHandling from './Settings-ImapFolderHandling'
 import SettingsManageKeys from './Settings-ManageKeys'
 
@@ -15,6 +16,7 @@ export function SettingsAdvanced({
   renderDeltaSwitch2: RenderDeltaSwitch2Type
   renderDTSettingSwitch: RenderDTSettingSwitchType
 }) {
+  const tx = window.static_translate
   return (
     <>
       <SettingsEncryption renderDeltaSwitch2={renderDeltaSwitch2} />
@@ -30,6 +32,13 @@ export function SettingsAdvanced({
       <br />
       <br />
       <SettingsTrayIcon
+        settingsStore={settingsStore}
+        renderDTSettingSwitch={renderDTSettingSwitch}
+      />
+      <br />
+      <br />
+      <H5>{tx('pref_experimental_features')}</H5>
+      <SettingsExperimentalFeatures
         settingsStore={settingsStore}
         renderDTSettingSwitch={renderDTSettingSwitch}
       />