+++ CHANGELOG.md
@@ -3,7 +3,10 @@
 ## [Unreleased][unreleased]
 
 ### Changed
+- Settings: Move Experimental Features into Advanced
+- Settings: Move Connectivity button down and give it an icon (like on android and ios)
 - update in-app help (2023-06-07)
+
 
 <a id="1_37_0"></a>